Add cancel_pending_call example; Add new actions; Add property callbacks.
It is now possible to set a callback function on any property of any
action (not all actions are supported yet). The callback will be called
(in a separate thread) once the property's value has been changed.
The (raw) value is passed to the callback as an argument.

BellButtonAction is new and is avaiable for intercom/bell routines.

Various other improvements and fixes have been implemented.